+++ example/password.dart
@@ -1,0 +1,2 @@
+String hiddenEmail = "your email address";
+String hiddenPassword = "your password";