+++ test/auth_test.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:auth070/auth.dart';
+
+void main() {
+  testWidgets('Auth test', (WidgetTester tester) async {
+//
+  });
+}